don't always clean in c.sh

+++ libcrux-ml-kem/c.sh
@@ -15,6 +15,7 @@ fi
 portable_only=0
 no_hacl=0
 no_charon=0
+clean=0
 
 # Parse command line arguments.
 all_args=("$@")
@@ -23,6 +24,7 @@ while [ $# -gt 0 ]; do
     -p | --portable) portable_only=1 ;;
     --no-hacl) no_hacl=1 ;;
     --no-charon) no_charon=1 ;;
+    -c | --clean) clean=1 ;;
     esac
     shift
 done
@@ -39,7 +41,7 @@ if [[ "$no_charon" = 0 ]]; then
     (cd ../libcrux-sha3 && RUSTFLAGS="--cfg eurydice" $CHARON_HOME/bin/charon --errors-as-warnings)
     if ! [[ -f ../libcrux_sha3.llbc ]]; then
       echo "😱😱😱 You are the victim of this bug: https://hacspec.zulipchat.com/#narrow/stream/433829-Circus/topic/charon.20declines.20to.20generate.20an.20llbc.20file"
-      echo "Suggestion: rm -rf ../target"
+      echo "Suggestion: rm -rf ../target or cargo clean"
       exit 1
     fi
     echo "Running charon (ml-kem) ..."
@@ -51,10 +53,14 @@ fi
 mkdir -p c
 cd c
 
-rm -rf *.c *.h
-# HAND_WRITTEN FILE
-git checkout libcrux_platform.c
-rm -rf internal/*.h
+# Clean only when requesting it.
+# Note that we can not extract for all platforms on any platform right now.
+# Make sure to keep files from other platforms.
+if [[ "$clean" = 1 ]]; then
+    rm -rf *.c *.h
+    rm -rf internal/*.h
+fi
+
 echo "Running eurydice ..."
 $EURYDICE_HOME/eurydice --config ../c.yaml ../../libcrux_ml_kem.llbc ../../libcrux_sha3.llbc
 cp $EURYDICE_HOME/include/eurydice_glue.h .
